% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_test.R
\name{auc_calculate}
\alias{auc_calculate}
\title{AUC probability and empirical p-value calculation}
\usage{
auc_calculate(x, y, N = 5000, rdm, p.output = FALSE)
}
\arguments{
\item{x}{numeric vector represents 1st sample}

\item{y}{numeric vector represents 2nd sample}

\item{N}{sampling times with replacement}

\item{rdm}{random seed}

\item{p.output}{indicator whether to return p-value}
}
\value{
AUC probability estimates (or with p-value)
}
\description{
AUC probability and empirical p-value calculation
}
\examples{
ecc.ratio =  sapply(1:nrow(otu),function(x){log_ratio(otu[x,which(meta$cariesfree==0),],version=1,epsilon=1)})
free.ratio =  sapply(1:nrow(otu),function(x){log_ratio(otu[x,which(meta$cariesfree==1),],version=1,epsilon=1)})
auc_prob_pvalue(ecc.ratio[[3]],free.ratio[[3]],rdm = i)

}
