% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_test.R
\name{cor.otu}
\alias{cor.otu}
\title{Correlation-test between log ratio and predictor}
\usage{
cor.otu(i, simu, all, xm = "t3c", method = "pearson")
}
\arguments{
\item{i}{ith genera or species specific taxonomic units which are used for testing}

\item{simu}{Operational taxonomic unit table}

\item{all}{the test we want to used for pre-analysis, \code{method = "LB"} represents logistics Beta test, while \code{method = "LN"} represents log Normal test}

\item{xm}{the main predictor we use}

\item{method}{a character string indicating which correlation coefficient is to be used for the test. One of "pearson", "kendall", or "spearman", can be abbreviated.}
}
\value{
Coefficient estimate and p-value of the main predictor
}
\description{
Correlation-test between log ratio and predictor
}
\examples{
all.ratio =  sapply(1:nrow(otu),function(x){log_ratio(otu[x,,],version=1,epsilon=1)})
cor.otu(1,otu,all.ratio)



}
