% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_test.R
\name{GMR.peaktest}
\alias{GMR.peaktest}
\title{Gaussian Mixture Regression for otu data}
\usage{
GMR.peaktest(i, simu, all, formula0 = "ratio ~ t3c + age", bindex, df = df)
}
\arguments{
\item{i}{ith genera or species specific taxonomic units which are used for testing}

\item{simu}{Operational taxonomic unit table}

\item{all}{the test we want to used for pre-analysis, \code{method = "LB"} represents logistics Beta test, while \code{method = "LN"} represents log Normal test}

\item{formula0}{a two-sided linear formula object describing all parts of the model. The main predictor \textbf{should be written in the first place} after tilde.}

\item{bindex}{the peak status of ith genera or species of otu, it has three values: \code{left},\code{right} and \code{both}, which refers to left-only peak, right-only peak and two peaks.}
}
\value{
Coefficient estimate and p-value of each peak
}
\description{
Gaussian Mixture Regression for otu data
}
\examples{
all.ratio =  sapply(1:nrow(otu),function(x){log_ratio(otu[x,,],version=1,epsilon=1)})
bindex = sapply(1:nrow(otu),function(i){ifelse(mean(otu[i,otu[i,,1]!=0,2] > 0) < 0.15,"left",ifelse(mean(otu[i,otu[i,,1]!=0,2] == 0)<0.15,"right","both"))})
GMR.peaktest(3,otu,all.ratio,formula0 = "ratio~t3c+age",bindex,df=df)

}
