% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{hist.comp}
\alias{hist.comp}
\title{Histogram of log ratio in disease and control group with density curves}
\usage{
\method{hist}{comp}(
  i,
  simu,
  disease,
  control,
  name = "ECC",
  position = "identity",
  binwidth = 0.1
)
}
\arguments{
\item{i}{ith genera or species specific taxonomic units which are used for testing}

\item{simu}{Operational taxonomic unit table}

\item{disease}{A list of log ratio in disease group}

\item{name}{Disease Name, showing in the legend}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment function.}

\item{binwidth}{The width of the bins. Can be specified as a numeric value or as a function that calculates width from unscaled x. Here, "unscaled x" refers to the original x values in the data, before application of any scale transformation.}

\item{conrol}{A list of log ratio in control group}
}
\value{
A ggplot showing the distribution of log ratio
}
\description{
Histogram of log ratio in disease and control group with density curves
}
\examples{
ecc.ratio =  sapply(1:nrow(otu),function(x){log_ratio(otu[x,which(meta$cariesfree==0),],version=1,epsilon=1)})
free.ratio =  sapply(1:nrow(otu),function(x){log_ratio(otu[x,which(meta$cariesfree==1),],version=1,epsilon=1)})
hist.comp(3,otu,ecc.ratio,free.ratio)

}
