% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{newdf}
\alias{newdf}
\title{Construct data-set including meta information and ratio}
\usage{
newdf(i, simu, all, df = df)
}
\arguments{
\item{i}{ith genera or species specific taxonomic units we used}

\item{simu}{Operational taxonomic unit table}

\item{all}{A list of log ratio for all the species/genes}

\item{df}{original meta data (including age, caries, race, reads and t3c)}
}
\value{
A data frame contaning ratio and which type it belongs to for every sample
}
\description{
Construct data-set including meta information and ratio
}
\examples{

ratio.otu = sapply(1:nrow(otu),function(x){log_ratio(otu[x,,],version=1,epsilon=1)})
newdf(1,otu,ratio.otu)




}
