% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{plot.total}
\alias{plot.total}
\title{Density curves of log ratio for the whole otu}
\usage{
plot.total(simu, id)
}
\arguments{
\item{simu}{Operational taxonomic unit table}

\item{id}{numeric vector, the row number of the genera/species we are interested}
}
\value{
A ggplot showing the distribution of log ratio
}
\description{
Density curves of log ratio for the whole otu
}
\examples{
plot.total(otu,1:109)

}
