% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preprocess_otu}
\alias{preprocess_otu}
\title{OTU preprocessing function}
\usage{
preprocess_otu(data)
}
\arguments{
\item{data}{A list of microbiome data containing the information of otu and meta data}
}
\value{
A list of microbiome data after preprocessing
}
\description{
OTU preprocessing function
}
\examples{

data(otu.data)
otu <- preprocess_otu(otu.data)



}
