% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_test.R
\name{pure.pretest}
\alias{pure.pretest}
\title{Pre-test for DNA and RNA}
\usage{
pure.pretest(
  i,
  otu,
  method = "LN",
  val = "DNA",
  xm = "caries_free",
  ranef = FALSE
)
}
\arguments{
\item{i}{ith genera or species specific taxonomic units which are used for testing}

\item{otu}{Operational taxonomic unit table}

\item{method}{the test we want to used for pre-analysis, \code{method = "LB"} represents logistics Beta test, while \code{method = "LN"} represents log Normal test}

\item{val}{DNA & RNA that needs to be tested}

\item{xm}{the main predictor we use (consider age and batch as covariates)}

\item{ranef}{whether we should put batch as a random effect}
}
\value{
Coefficient estimate and p-value of the main predictor
}
\description{
Pre-test for DNA and RNA
}
\examples{

pure.pretest(1,otu)



}
