% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ratio_factor}
\alias{ratio_factor}
\title{Calculate ratio and type}
\usage{
ratio_factor(species_data, epsilon = 1e-06)
}
\arguments{
\item{species_data}{p*2 matrix, p represents the samples, DNA and RNA columns}

\item{epsilon}{the constant we add in both denominator and numerator}
}
\value{
A data frame contaning ratio and which type it belongs to for every sample
}
\description{
Calculate ratio and type
}
\examples{

ratio_factor(otu[1,,])




}
