% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{scatter.total}
\alias{scatter.total}
\title{Scatterplot of log DNA versus log RNA}
\usage{
scatter.total(simu, id)
}
\arguments{
\item{simu}{Operational taxonomic unit table}

\item{id}{numeric vector, the row number of the genera/species we are interested}
}
\value{
A ggplot showing the scatterplot of log DNA versus log RNA
}
\description{
Scatterplot of log DNA versus log RNA
}
\examples{
scatter.total(otu,1:109)

}
