% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_test.R
\name{ttest}
\alias{ttest}
\title{T-test for comparing the element in two log ratio list}
\usage{
ttest(i, z, w)
}
\arguments{
\item{i}{ith species/genes which are used for comparison}

\item{z}{the first samples}

\item{w}{the second samples}
}
\value{
Vector of t-test p-value
}
\description{
T-test for comparing the element in two log ratio list
}
\examples{
ecc.ratio =  sapply(1:nrow(otu),function(x){log_ratio(otu[x,which(meta$cariesfree==0),],version=1,epsilon=1)})
free.ratio =  sapply(1:nrow(otu),function(x){log_ratio(otu[x,which(meta$cariesfree==1),],version=1,epsilon=1)})
ttest(1,ecc.ratio,free.ratio)




}
