% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_test.R
\name{ratio.test}
\alias{ratio.test}
\title{Ratio level test}
\usage{
ratio.test(i, simu, all, xm = "t3c", reg = "linear", ranef = TRUE, df = df)
}
\arguments{
\item{i}{ith genera or species specific taxonomic units which are used for testing}

\item{simu}{Operational taxonomic unit table}

\item{all}{the test we want to used for pre-analysis, \code{method = "LB"} represents logistics Beta test, while \code{method = "LN"} represents log Normal test}

\item{xm}{the main predictor we use}

\item{reg}{kind of regression that we want to see the p-value of it, it has \code{linear} and \code{logistics} two choices.}

\item{ranef}{whether we should regard batch as random effect}

\item{df}{meta data we use}
}
\value{
Coefficient estimate and p-value of the main predictor
}
\description{
Ratio level test
}
\examples{
all.ratio =  sapply(1:nrow(otu),function(x){log_ratio(otu[x,,],version=1,epsilon=1)})
ratio.test(1,otu,all.ratio,reg="linear",df=df)

}
